+++ src/app/containers/Picture/constants.js
