+++ src/app/containers/Post/constants.js
